move GetXLogRecordData from test to testtools
test/delta_file_manager_test.go:GetXLogRecordData -> testtools/util.go:GetXLogRecordData
