+++ test/block_locations_util_test.go
@@ -4,11 +4,12 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/wal-g/wal-g/internal"
 	"github.com/wal-g/wal-g/internal/walparser"
+	"github.com/wal-g/wal-g/testtools"
 	"testing"
 )
 
 func TestExtractBlockLocations(t *testing.T) {
-	record, _ := GetXLogRecordData()
+	record, _ := testtools.GetXLogRecordData()
 	expectedLocations := []walparser.BlockLocation{record.Blocks[0].Header.BlockLocation}
 	actualLocations := internal.ExtractBlockLocations([]walparser.XLogRecord{record})
 	assert.Equal(t, expectedLocations, actualLocations)
